работа над ошибками
