+++ lot.php
@@ -1,8 +1,5 @@
 <?php
-session_start();
-require_once ('helpers.php');
-require_once ('bd.php');
-
+require_once('init.php');
 if (!$con) {
     $content = include_template('404.php');
 }
@@ -19,10 +16,10 @@ else {
             $ads_lot = get_ads_lot($lot_id, $con);
             $ads_lot['lot_timer'] = get_date($ads_lot['calculation_date'])['is_finishing'];
             $ads_lot['calculation_date'] = get_date($ads_lot['calculation_date'])['times'];
-            $content = include_template('model_lot.php', ['ads_lot' => $ads_lot]);
+            $main_content = include_template('model_lot_template.php', ['ads_lot' => $ads_lot]);
+            $content = include_template('other_layout.php', ['content' => $main_content, 'categories' => $categories, 'user' => $_SESSION, 'title' => 'Добавление лота']);
         }
     }
 }
-
 print($content);
