+++ templates/login_template.php
@@ -1,0 +1,15 @@
+<form class="form container<?= empty($errors) ? "" : " form--invalid"; ?>" action="login.php" method="post"> <!-- form--invalid -->
+    <h2>Вход</h2>
+    <div class="form__item<?= empty($errors['email']) ? "" : " form__item--invalid"; ?>"> <!-- form__item--invalid -->
+        <label for="email">E-mail <sup>*</sup></label>
+        <input id="email" type="text" name="email" placeholder="Введите e-mail" value="<?= get_post_val($login['email']); ?>">
+        <span class="form__error"><?= $errors['email'] ?? ""; ?></span>
+    </div>
+    <div class="form__item form__item--last<?= empty($errors['password'] || $errors['password_verification']) ? "" : " form__item--invalid"; ?>">
+        <label for="password">Пароль <sup>*</sup></label>
+        <input id="password" type="password" name="password" placeholder="Введите пароль">
+        <span class="form__error"><?= $errors['password'] ?? ""; ?></span>
+        <span class="form__error"><?= $errors['password_verification'] ?? ""; ?></span>
+    </div>
+    <button type="submit" class="button">Войти</button>
+</form>
